programs/ie-baseline/requirements.txt
tqdm
tensorboard<2.4
pytorch